Remove AsyncTask
