+++ android-examples/app/src/main/res/layout/activity_main.xml
@@ -8,10 +8,12 @@
 
 
     <view
+        android:id="@+id/id_scatterview"
         class="com.example.androidDl4jClassifier.ScatterView"
         android:layout_width="fill_parent"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
